fix the pathing for the new file structure and add a sharedStyles.js
