+++ sharedStyles.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 
-const styles = StyleSheet.create({
+const sharedStyles = StyleSheet.create({
   container: {
     flex: 1,
     backgroundColor: '#fff',
@@ -42,4 +42,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default styles;
+export default sharedStyles;
